Create README.md
This project helps one to visualize a sorting algorithm. Each element of the array is displayed as a bar. The operations are colour coded such that:

Red - Swap
Blue - Comparison
Green - Element is in sorted position
It compares the time taken by the different algorithm for sorting the array.

Technologies
HTML
CSS
Javascript (p5.js)